Move blades logic form API to service layer
This is a requirement for further service composition.
