+++ ruggedpod_api/services/network.py
@@ -16,6 +16,9 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import re
+import socket
+import fcntl
+import struct
 
 from ruggedpod_api.common import exception
 from ruggedpod_api.services import dhcp, utils
@@ -30,6 +33,15 @@ cache_opts = {
 }
 
 cache = CacheManager(**parse_cache_config_options(cache_opts))
+
+
+def get_iface_ip_address(ifname):
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    return socket.inet_ntoa(fcntl.ioctl(
+        s.fileno(),
+        0x8915,  # SIOCGIFADDR
+        struct.pack('256s', ifname[:15])
+    )[20:24])
 
 
 def normalize_mac_address(mac):
